recurse into imported documents
