+++ src/main/scala/wdlTools/FetchURL.scala
@@ -1,0 +1,7 @@
+package wdlTools
+
+import ConcreteSyntax._
+
+object FetchURL {
+  def apply(url: URL): String = ???
+}
